♻️ Refactor: history, draw, visualizer 컴포넌트 분리
